# openapi-ci-api.yaml
openapi: 3.0.3
info:
  title: CCExtractor CI System API
  version: 1.2.0
  description: |
    Security-hardened JSON-only REST API for the CCExtractor CI/sample platform.
    Designed for AI agents and CI automation. Enforces scoped Bearer token auth,
    strict input validation, rate limiting on all routes, and safe defaults
    throughout. No browser sessions, no HTML, no implicit permissions.

    **Authentication:** All endpoints require bearer token authentication unless
    explicitly marked with `security: []` (only /system/health and POST /auth/tokens).

    **Rate-limit headers:** Every response includes `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` headers. These are modeled
    explicitly on the 429 response for brevity; they are present on all responses
    regardless of status code.

  contact:
    name: CCExtractor Development
    url: https://github.com/CCExtractor/sample-platform
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://sampleplatform.ccextractor.org/api/v1
    description: Production

#
# Global security: all endpoints require auth
# unless explicitly overridden with security: []
#
security:
  - bearerAuth: []
...
paths:
...
  /runs/{run_id}/samples:
    get:
      tags: [Samples]
      summary: List regression test results in a run
      operationId: listRunSamples
      description: >
        Returns one entry per regression test result, not one per unique media
        file. A single media sample may yield multiple entries if it has
        multiple regression tests (different command flags).
        sample_progress in the legacy JSON endpoint is len(test.results) over
        total regression tests; it does not reflect multi-output completeness.
      security:
        - bearerAuth: []
      x-required-scope: runs:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: status
          in: query
          schema:
            type: string
            enum: [pass, fail, skipped, missing_output, running, not_started]
        - name: name
          in: query
          schema:
            type: string
            maxLength: 100
        - name: tag
          in: query
          schema:
            type: string
            maxLength: 50
        - name: category
          in: query
          schema:
            type: string
            maxLength: 50
      responses:
        "200":
          description: Paginated regression test results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/RunSample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /runs/{run_id}/samples/{regression_test_id}:
    get:
      tags: [Samples]
      summary: Get full details for a regression test result in a run
      operationId: getRunSample
      description: >
        Returns the result for a specific regression test within a run.
        Note: the path parameter is regression_test_id, not a media sample ID.
        A single media sample may have multiple regression tests.
      security:
        - bearerAuth: []
      x-required-scope: runs:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/RegressionTestId"
      responses:
        "200":
          description: Regression test result details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunSample"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /samples:
    get:
      tags: [Samples]
      summary: List all known media samples
      operationId: listSamples
      description: >
        Returns paginated media sample metadata. Samples are the original
        media files uploaded for regression testing.
      security:
        - bearerAuth: []
      x-required-scope: runs:read
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: status
          in: query
          description: >
            Derived from linked regression tests. The sample table itself has
            no quarantine state; active/inactive reflects whether any active
            regression tests reference the sample.
          schema:
            type: string
            enum: [active, inactive]
        - name: name
          in: query
          schema:
            type: string
            maxLength: 100
        - name: tag
          in: query
          schema:
            type: string
            maxLength: 50
        - name: sha256
          in: query
          schema:
            type: string
            pattern: '^[a-fA-F0-9]{64}$'
        - name: extension
          in: query
          schema:
            type: string
            maxLength: 10
            pattern: '^[a-zA-Z0-9]+$'
      responses:
        "200":
          description: Paginated media samples
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Sample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /samples/{sample_id}:
    get:
      tags: [Samples]
      summary: Get media sample metadata
      operationId: getSample
      security:
        - bearerAuth: []
      x-required-scope: runs:read
      parameters:
        - $ref: "#/components/parameters/SampleId"
      responses:
        "200":
          description: Media sample metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sample"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /samples/{sample_id}/history:
    get:
      tags: [Samples]
      summary: Get regression test result history for a sample across runs
      operationId: getSampleHistory
      description: >
        Use failure_signature for flake detection: a stable signature across
        multiple runs on different commits indicates a genuine regression,
        not infrastructure noise.
      security:
        - bearerAuth: []
      x-required-scope: runs:read
      parameters:
        - $ref: "#/components/parameters/SampleId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/RunStatus"
        - $ref: "#/components/parameters/Branch"
        - $ref: "#/components/parameters/Platform"
        - $ref: "#/components/parameters/CreatedAfter"
        - $ref: "#/components/parameters/CreatedBefore"
      responses:
        "200":
          description: Paginated sample history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/SampleHistoryEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /regression-tests:
    get:
      tags: [Samples]
      summary: List regression test definitions
      operationId: listRegressionTests
      description: >
        The active filter must be applied explicitly. When no custom set is
        defined, all regression tests are returned — including inactive ones.
      security:
        - bearerAuth: []
      x-required-scope: runs:read
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: active
          in: query
          schema:
            type: boolean
            default: true
        - name: category
          in: query
          schema:
            type: string
            maxLength: 50
        - name: tag
          in: query
          schema:
            type: string
            maxLength: 50
        - name: sample_id
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Paginated regression test definitions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/RegressionTest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"

  # RESULTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Opaque server-side API token. Obtain via POST /auth/tokens.
        The CI worker token used by /ci/progress-reporter is a separate
        secret and is NOT valid here. Never use browser session cookies
        for API clients.

  # HEADERS
...
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return (1–100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      description: Number of results to skip for pagination
      schema:
        type: integer
        minimum: 0
        maximum: 2147483647
        default: 0
...
    RunId:
      name: run_id
      in: path
      required: true
      description: Numeric run ID
      schema:
        type: integer
        minimum: 1
...
    RegressionTestId:
      name: regression_test_id
      in: path
      required: true
      description: Numeric regression test ID (not the same as media sample ID)
      schema:
        type: integer
        minimum: 1
...
  responses:
    BadRequest:
      description: Request body or query parameters failed schema validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: validation_error
            message: Request failed schema validation.
            details:
              fields:
                commit_sha: Must match pattern ^[a-fA-F0-9]{40}$
                platform: Must be one of [linux, windows]
    Unauthorized:
      description: Missing, expired, or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: unauthorized
            message: Bearer token is missing, expired, or invalid.
            details: {}
    Forbidden:
      description: Token is valid but lacks the required scope or role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: forbidden
            message: Token does not have the required scope for this operation.
            details:
              required_scope: runs:write
              token_scopes: [runs:read, results:read]
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: not_found
            message: Run 9317 not found.
            details:
              resource: run
              id: 9317
...
    RateLimited:
      description: Too many requests. Retry after the indicated number of seconds.
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 30
        X-RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: rate_limited
            message: Rate limit exceeded. Retry after 30 seconds.
            details:
              retry_after: 30
              limit: 120
              window: 60s
    Error:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # SCHEMAS
  schemas:
    Page:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          description: >
            Result items. The concrete type is defined by allOf composition
            in each endpoint response.
          items: {}
        pagination:
          type: object
          required: [limit, offset, total]
          properties:
            limit:
              type: integer
              minimum: 1
            offset:
              type: integer
              minimum: 0
            total:
              type: integer
              minimum: -1
              nullable: true
              description: >
                Total matching records. Null if count was not computed for this request.
                Pass ?count=true to force computation.
            next_offset:
              type: integer
              minimum: 0
              nullable: true
            truncated:
              type: boolean
              description: >
                Present and true when the result set was capped by an
                internal safety limit (e.g. status-filter on runs). When
                true, total may undercount the real number of matches.
...
    RunSample:
      type: object
      required: [regression_test_id, sample_id, status]
      properties:
        regression_test_id:
          type: integer
          minimum: 1
        sample_id:
          type: integer
          minimum: 1
          nullable: true
        sample_name:
          type: string
          maxLength: 255
          nullable: true
        categories:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 100
          description: Category labels from the regression test definition.
        command:
          type: string
          maxLength: 500
          nullable: true
        status:
          type: string
          enum: [pass, fail, skipped, missing_output, running, not_started]
          description: >
            Computed from TestResult, TestResultFile, expected exit code,
            and multiple acceptable baselines. Not a stored column.
        runtime_ms:
          type: integer
          minimum: 0
          nullable: true
        exit_code:
          type: integer
          nullable: true
        expected_rc:
          type: integer
          nullable: true
          description: Expected return code for this regression test.
        outputs:
          type: array
          maxItems: 20
          description: >
            One entry per expected output file.
            got=null in the DB means output matched expected; no actual file
            is stored. The dummy (-1,-1,-1,'','error') row is translated to
            status=missing_output and is never exposed here.
          items:
            type: object
            required: [output_id, filename, status]
            additionalProperties: false
            properties:
              output_id:
                type: integer
                minimum: 1
              filename:
                type: string
                maxLength: 255
              status:
                type: string
                enum: [pass, fail, missing_output, missing_expected]
                description: >
                  pass = actual identical to expected.
                  fail = actual differs from expected.
                  missing_output = test produced no output.
                  missing_expected = no expected baseline exists.
...